% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_class}
\alias{check_class}
\title{check if an object is of a certain class}
\usage{
check_class(x, classname)
}
\arguments{
\item{x}{object}

\item{classname}{character}
}
\value{
nothing, emit error if check not satisfied
}
\description{
check if an object is of a certain class
}
\keyword{internal}
