% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_df}
\alias{check_df}
\title{validate degrees of freedom.}
\usage{
check_df(x, max.value)
}
\arguments{
\item{x}{expect a single number}

\item{max.value}{numeric, maximal value allowed for x}
}
\value{
integer derived from x
}
\description{
validate degrees of freedom.
}
\keyword{internal}
