% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_lambda}
\alias{check_lambda}
\title{validate lambda. They must be numeric, finite, sorted with unique value}
\usage{
check_lambda(x, pmax)
}
\arguments{
\item{x}{vector of lambda values}

\item{pmax}{numeric, maximal pvalue}
}
\value{
numeric vector of sorted unique x, or stop if x does not satisfy criteria
}
\description{
validate lambda. They must be numeric, finite, sorted with unique value
}
\keyword{internal}
