% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_p}
\alias{check_p}
\title{validate pvalues. They must be finite, in range [0,1]}
\usage{
check_p(p)
}
\arguments{
\item{p}{numeric vector of p-values}
}
\value{
numeric vector of length 2 with (min(p), max(p))
}
\description{
validate pvalues. They must be finite, in range [0,1]
}
\keyword{internal}
