% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_X}
\alias{check_X}
\title{validate matrix of covariates. It must be compatible with a vector of pvalues}
\usage{
check_X(X, p)
}
\arguments{
\item{X}{vector or matrix of covariates}

\item{p}{vector of p-values}
}
\value{
matrix
}
\description{
validate matrix of covariates. It must be compatible with a vector of pvalues
}
\keyword{internal}
