% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compose.stats}
\alias{compose.stats}
\title{Compose a two line report about a numeric vector}
\usage{
compose.stats(v)
}
\arguments{
\item{v}{numeric vector}
}
\value{
vector with two strings a header line and a data line
}
\description{
Compose a two line report about a numeric vector
}
\keyword{internal}
