% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{compound.message}
\alias{compound.message}
\title{Compose and output a compound message and output}
\usage{
compound.message(x, components = c("call", "lambda", "X", "pi0", "hits"))
}
\arguments{
\item{x}{list object of type lm_qvalue or lm_pi0 (not checked)}

\item{components}{character vector, identifiers suggesting what to include in output}
}
\value{
character vector
}
\description{
Compose and output a compound message and output
}
\keyword{internal}
