% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estPi0Reg.R
\name{estPi0Reg}
\alias{estPi0Reg}
\title{Estimate pi0(x)}
\usage{
estPi0Reg(p, lambda = seq(0.05, 0.95, 0.05), X, smooth.df = 3)
}
\arguments{
\item{p}{Numerical vector of p-values}

\item{lambda}{Numerical vector of thresholds. Must be in [0,1).}

\item{X}{Design matrix (one test per row, one variable per column). Do not include the intercept.}

\item{smooth.df}{Number of degrees of freedom when estimating pi0(x) with a smoother.}
}
\value{
pi0 Numerical vector of smoothed estimate of pi0(x). The length is the number of rows in X.

pi0.lambda Numerical matrix of estimated pi0 for each value of lambda. The number of columns is the number of tests, the number of rows is the length of lambda.

lambda Vector of the values of lambda used in calculating pi0.lambda

pi0.smooth Matrix of fitted values from the smoother fit to the pi0(x) estimates at each value of lambda (same number of rows and columns as pi0.lambda)
}
\description{
Estimate pi0(x)
}

