% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastspline.R
\name{fast.spline}
\alias{fast.spline}
\title{Fast smoothing splines}
\usage{
fast.spline(x, y = NULL, df = 3, tol = 4.5e-07)
}
\arguments{
\item{x}{numeric vector, must have sorted, unique, finite values}

\item{y}{numeric vector, must have finite values}

\item{df}{integer, number of degrees of freedom}

\item{tol}{numeric, tolerance that determines when smoothing stops.}
}
\value{
numeric vector with smoothed y values
}
\description{
Smoothing is computed using the same code as smooth.spline from the stats
package. This version can run substantially faster because it does not
perform validation checks on the inputs. The inputs must be prepared
outside of the function. This setup can be a benefit if smoothing
is performed many times with the same values of x.
}
\details{
This version also supports a very limited number of arguments. The
behavior of this function follows the default settings for smooth.spline.

Note: this function uses a function internal to the stats package to
perform its work. In principle, this means fast.spline can break
if that function from stats changes. However, that function from stats
has remained fixed for many years (now 2018) and is thus not likely to
change. If that happens, the results from fast.spline can be reproduced
via 'smooth.spline(x, y, df=3, tol=4.5e-7)$y'
}
\keyword{internal}
