% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pi0.R
\name{fit_linear}
\alias{fit_linear}
\title{Fit response values using a linear regression}
\usage{
fit_linear(y, X)
}
\arguments{
\item{y}{numeric vector}

\item{X}{numeric matrix (covariates)}
}
\value{
numeric vector
}
\description{
(This could be implemnted via glm(... family=gaussian) but the
implementation with lsfit is much faster
}
\keyword{internal}
