% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pi0.R
\name{fit_logistic}
\alias{fit_logistic}
\title{Fit response values using a binomial/logit model}
\usage{
fit_logistic(y, X)
}
\arguments{
\item{y}{numeric vector}

\item{X}{numeric matrix (covariates)}
}
\value{
numeric vector
}
\description{
Fit response values using a binomial/logit model
}
\keyword{internal}
