% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{get.components}
\alias{get.components}
\title{Get a set intersection, but when second set is null default to first set}
\usage{
get.components(supported, requested)
}
\arguments{
\item{supported}{vector of supported feature names}

\item{requested}{vector of requested feature names}
}
\value{
character vector with an intersection, or all supported features
}
\description{
This is meant to identify a subset of supported features that are requested
}
\keyword{internal}
