% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{hist.lm_pi0}
\alias{hist.lm_pi0}
\title{Display a histogram of pi0 estimates}
\usage{
\method{hist}{lm_pi0}(
  x,
  breaks = 31,
  xlab = "pi0 estimate",
  ylab = "Counts",
  main = "",
  css = "swfdr",
  cssclass = NULL,
  lines = 10,
  ...
)
}
\arguments{
\item{x}{object of class lm_pi0}

\item{breaks}{integer or vector, position of histogram breakpoints}

\item{xlab}{character, label for x-axis}

\item{ylab}{character, label for y-axis}

\item{main}{character, label for title}

\item{lines}{integer, bob}

\item{...}{other arguments}

\item{Rcss}{style object of class Rcss}

\item{Rcssclass}{character, style class}
}
\description{
Display a histogram of pi0 estimates
}
\examples{
# generate a lm_pi0 object
pValues <- rbeta(1000, 0.5, 1)
X <- pValues + rnorm(1000, 0, 0.5)
pi0 <- lm_pi0(pValues, X=X)
# draw histogram of pi0 estimates
hist(pi0)

}
