% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfdr.R
\name{lfdr}
\alias{lfdr}
\title{Estimation of local false discovery rate}
\usage{
lfdr(p, pi0, eps = 1e-08, adj = 1.5)
}
\arguments{
\item{p}{numeric vector of p-values}

\item{pi0}{numeric vector of pi0 estimates (same length as p)}

\item{eps}{numeric, regularization of p-values, same effect as in qvalue::lfdr}

\item{adj}{numeric, adjustment of smoothing bandwidth in density estimte, same effect
as in qvalue::lfdr}
}
\value{
numeric vector of local fdr values, truncated into interval [0, 1]
}
\description{
Local FDR is a comparison of a p-value distribution with a flat/uniform distribution.
This first half of this implementation is similar to the code for lfdr in package qvalue.
The second half deviates from qvalue::lfdr in that it avoids comparing the right tail
of the p-values with dnorm. The overall effect is to avoid producing pathologically low
estimates of lfdr for p-values close to unity. This pathological behavior is avoided
in qvalue::lfdr by using argument monotone=TRUE; here, the monotone transformation is
omitted to allow adjustment with a vector of pi0, and thus the pathological behavior
is handled through the block containing dnorm_flat.
}
