% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pi0.R
\name{lm_pi0_2}
\alias{lm_pi0_2}
\title{Another version of lm_pi0}
\usage{
lm_pi0_2(p, lambda = seq(0.05, 0.95, 0.05), X, type = c("logistic",
  "linear"))
}
\arguments{
\item{p}{numeric vector}

\item{lambda}{numeric vector}

\item{X}{matrix of covariates}

\item{type}{character}

\item{smooth}{character}
}
\value{
list with pi0, pi0.lambda, lambda
}
\description{
Another version of lm_pi0
}
