% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{optimize.multiplier}
\alias{optimize.multiplier}
\title{recursive internal function to find an optimal value of lambda
so that trace(X+lambda*omega) = target}
\usage{
optimize.multiplier(X, omega, target = 5, interval = c(1e-06, 1, 2),
  values = c(NA, NA, NA), tol = 1e-04)
}
\arguments{
\item{X}{matrix}

\item{omega}{matrix}

\item{target}{numeric, target for trace(X+lambda*omega)}

\item{interval}{numeric vector of length 2, current range for lambda (internal use)}

\item{values}{numeric vector of length 2, current target estimates (internal use)}

\item{tol}{numeric, numerical tolerance, does not need to be very small}
}
\value{
value of lagrange multiplier that bring criterion close to its target
}
\description{
This implementation relies on the knowledge that a larger lagrange multiplier
leads to lower trace(X+lambda*omega).
}
\details{
The implementation starts in an interval (0, 1, 2) and the expands the interval
or zooms in to find an
}
\keyword{internal}
