% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lm_qvalue}
\alias{plot.lm_qvalue}
\title{Display a scatter-plot compring p-values and q-values}
\usage{
\method{plot}{lm_qvalue}(
  x,
  var.x = c("pvalues", "qvalues.raw"),
  log = "xy",
  threshold = 0.05,
  xlab = NULL,
  ylab = NULL,
  main = "",
  css = "swfdr",
  cssclass = NULL,
  ...
)
}
\arguments{
\item{x}{object of class lm_pi0}

\item{var.x}{characer, variable to show on the x-axis}

\item{log}{character, use 'xy' to display both axes on logarithmic scale}

\item{threshold}{numeric, threshold to call significance is displayed using
lines, set to NULL to avoid showing the threshold lines}

\item{xlab}{character, label for x-axis}

\item{ylab}{character, label for y-axis}

\item{main}{character, label for plot title}

\item{css}{object of class Rcss}

\item{cssclass}{character, style class}

\item{...}{other arguments, passed to points()}
}
\description{
Display a scatter-plot compring p-values and q-values
}
\examples{
# generate a lm_pi0 object
pValues <- rbeta(1000, 0.5, 1)
X <- pValues + rnorm(1000, 0, 0.5)
qValues <- lm_qvalue(pValues, X=X)
# draw a comparison of p-values and q-values
plot(qValues)
plot(qValues, threshold=NA)

}
\keyword{internal}
