% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lm_pi0}
\alias{print.lm_pi0}
\title{Display a summary of object lm_pi0}
\usage{
\method{print}{lm_pi0}(x, ...)
}
\arguments{
\item{x}{object of class lm_pi0}

\item{...}{other arguments ignored}
}
\description{
Display a summary of object lm_pi0
}
\keyword{internal}
