% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.lm_qvalue}
\alias{print.lm_qvalue}
\title{Display a summary of an lm_qvalue object}
\usage{
\method{print}{lm_qvalue}(x, ...)
}
\arguments{
\item{x}{lm_qvalue object}

\item{...}{ignored}
}
\description{
Display a summary of an lm_qvalue object
}
\keyword{internal}
