% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_pi0.R
\name{regularize.interval}
\alias{regularize.interval}
\title{Force a set of values in a regular interval}
\usage{
regularize.interval(x, interval = c(0, 1))
}
\arguments{
\item{x}{numeric vector or matrix}

\item{interval}{numeric vector of length 2 with min/max values}
}
\value{
same object like x, with values truncated by [0,1]
}
\description{
(This is a simpler/faster implementation than with ifelse)
}
\keyword{internal}
