% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{smooth.spline.pi0}
\alias{smooth.spline.pi0}
\title{Fit smoothing splines to several curves}
\usage{
smooth.spline.pi0(x, ymat, df = 3)
}
\arguments{
\item{x}{numeric vector, position of knots}

\item{ymat}{numeric matrix, ncol(ymat) should match length(x)}
}
\value{
list with components pi0.smooth and pi0
}
\description{
This implementation uses smooth.spline. It gives a slight optimization
by not reporting original data (keep.data=FALSE) and by setting numeric
tolerance to a fixed number (tol=1e-7). In the latter, the precise number
does not matter, but it is important that it is not computed from x
at each iteration in the loop.
}
\keyword{internal}
