% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.lm_pi0}
\alias{summary.lm_pi0}
\title{Display a summary of object lm_pi0}
\usage{
\method{summary}{lm_pi0}(object, ...)
}
\arguments{
\item{object}{object of class lm_pi0}

\item{...}{other arguments ignored}
}
\description{
(This is the same as print.lm_pi0)
}
\keyword{internal}
