% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.lm_qvalue}
\alias{summary.lm_qvalue}
\title{Display a summary of object lm_qvalue}
\usage{
\method{summary}{lm_qvalue}(object, ...)
}
\arguments{
\item{object}{object of class lm_qvalue}

\item{...}{other arguments ignored}
}
\description{
(This is the same as print.lm_qvalue)
}
\keyword{internal}
