% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{unit.spline.last}
\alias{unit.spline.last}
\title{Fit smoothing unit splines to several series and obtain last fitted value}
\usage{
unit.spline.last(x, ymat, df = 3)
}
\arguments{
\item{x}{numeric vector, position of knots}

\item{ymat}{numeric matrix, ncol(ymat) should match length(x)}
}
\value{
numeric vector
}
\description{
In contrast to smooth.spline.last, this function uses the bs() bases
and uses boundary knots at x=(0, 1)
}
\keyword{internal}
