% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{unit.spline.matrix}
\alias{unit.spline.matrix}
\title{Construct a transformation matrix for a unit spline}
\usage{
unit.spline.matrix(x, df = 3)
}
\arguments{
\item{x}{numeric vector, position of internal knots within [0,1]}
}
\value{
matrix of dimension c(length(x), length(x))
}
\description{
This function uses library::bs to construct a b-spline basis
from knots, using [0,1] as the boundary knots.
}
\keyword{internal}
