% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{unit.spline.pi0}
\alias{unit.spline.pi0}
\title{Fit smoothing unit splines to several curves}
\usage{
unit.spline.pi0(x, ymat, df = 3)
}
\arguments{
\item{x}{numeric vector, position of knots}

\item{ymat}{numeric matrix, ncol(ymat) should match length(x)}
}
\value{
list with component pi0 (numeric vector)
}
\description{
In contrast to smooth.spline.pi0, this function uses the bs() bases
and uses boundary knots at x=(0, 1)
}
\keyword{internal}
