% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_pi0.R
\name{unit.spline}
\alias{unit.spline}
\title{Fit smoothing unit spline to one curve}
\usage{
unit.spline(x, y, df = 3)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
vector of same length as y
}
\description{
Fit smoothing unit spline to one curve
}
\keyword{internal}
