% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{v2s}
\alias{v2s}
\title{Helper to create a single string by concatenating items from a vector}
\usage{
v2s(x, width = 8)
}
\arguments{
\item{x}{vector of things}

\item{width}{vector of character widths for each item in x}
}
\value{
a single string
}
\description{
Helper to create a single string by concatenating items from a vector
}
\keyword{internal}
