% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSwfdr.R
\name{calculateSwfdr}
\alias{calculateSwfdr}
\title{Calculate the science-wise FDR}
\usage{
calculateSwfdr(pValues, truncated, rounded, pi0 = 0.5, alpha = 1,
  beta = 50, numEmIterations = 100)
}
\arguments{
\item{pValues}{Numerical vector of p-values}

\item{truncated}{Vector of 0s and 1s with indices corresponding to those in pValues; 1 indicates that the p-values is truncated, 0 that it is not truncated}

\item{rounded}{Vector of 0s and 1s with indices corresponding to those in pValues; 1 indicates that the p-values is rounded, 0 that it is not rounded}

\item{pi0}{Initial prior probability that a hypothesis is null}

\item{alpha}{Initial value of parameter alpha from Beta(alpha, beta) true positive distribution}

\item{beta}{Initial value of parameter beta from Beta(alpha, beta) true positive distribution}

\item{numEmIterations}{The number of EM iterations (default is 100)}
}
\value{
pi0 Final value of prior probability - estimated from EM - that a hypothesis is null

alpha Final value of parameter alpha - estimated from EM - from Beta(alpha, beta) true positive distribution

beta Final value of parameter beta - estimated from EM - from Beta(alpha, beta) true positive distribution

z Indicator vector of 0s and 1s - estimated from EM - with 1s indicating p-values from the null distribution

n0 Expected number of null p-values - estimated from EM - between certain cutpoints (0.005, 0.015, 0.025, 0.035, 0.045, 0.051)

n Number of p-values between certain cutpoints (0.005, 0.015, 0.025, 0.035, 0.045, 0.051)
}
\description{
Calculate the science-wise FDR
}

