% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_number_decimals.R
\name{get_number_decimals}
\alias{get_number_decimals}
\title{Get number of decimals for any vector of numbers (i.e. return total number of digits after decimal point)}
\usage{
get_number_decimals(x)
}
\arguments{
\item{x}{Numerical vector}
}
\value{
Vector giving the number of decimals for each element in x
}
\description{
Get number of decimals for any vector of numbers (i.e. return total number of digits after decimal point)
}
\examples{
get_number_decimals(c(1,2))
get_number_decimals(c(0.0006, 0.0750, 0.0420, 0.0031, 0.0001, 0.0100))
get_number_decimals(c(6.5*10^-5, 0.0100))
get_number_decimals(c(10.5,6.57*10^-5))
get_number_decimals(c(-1000.05,-6.57*10^-5))

}
