% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_qvalue.R
\name{lm_qvalue}
\alias{lm_qvalue}
\title{Compute qvalues taking into account a matrix of covariates}
\usage{
lm_qvalue(
  p,
  X,
  pfdr = FALSE,
  pi0 = NULL,
  smoothing = c("unit.spline", "smooth.spline"),
  ...
)
}
\arguments{
\item{p}{numeric vector of p-values}

\item{X}{matrix of covariates (can be missing if pi0 is specified instead)}

\item{pfdr}{logical, making estimates robust for small p-values and a small
sample size}

\item{pi0}{list with pi0 estimates from lm_pi0}

\item{smoothing}{character, type of smoothing used to fit pi0. Note the default
in this function is different than in lm_pi0.}

\item{...}{other parameters (passed on to lm_pi0 if pi0 is not provided)}
}
\value{
list
}
\description{
The recipe for turning pvalues into qvalues is adapted from package
'qvalue' and articles by Storey, Tibshirani, Taylor, Siegmund.
}
